requirements.txt
# =============================================================================
# ESG Data Foundry — Python dependencies (Gods Finger v2)
# =============================================================================

# ── LangGraph / LangChain ─────────────────────────────────────────────────
langgraph>=0.2.0,<0.3
langchain>=0.2.0,<0.3
langchain-openai>=0.1.0,<0.2
langchain-community>=0.2.0,<0.3

# ── OpenAI (Judge + Embeddings + fallback) ────────────────────────────────
openai>=1.40.0,<2.0

# ── Database ──────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
sqlalchemy>=2.0.30,<3.0
pgvector>=0.3.0

# ── Configuration ─────────────────────────────────────────────────────────
pydantic>=2.7.0,<3.0
pydantic-settings>=2.3.0,<3.0
python-dotenv>=1.0.1

# ── Resilience ────────────────────────────────────────────────────────────
tenacity>=8.3.0,<9.0

# ── PDF parsing ───────────────────────────────────────────────────────────
llama-parse>=0.4.9          # LlamaParse cloud — najlepsza jakość tabel ($0.003/str)
pdf2image>=1.17.0           # OpenAI Vision fallback (wymaga poppler-utils)
pymupdf>=1.24.0             # Local PDF parser (darmowy, bez limitów)
Pillow>=10.3.0

# ── DOCX parsing ──────────────────────────────────────────────────────────
python-docx>=1.1.0          # DOCX/DOC → Markdown (lokalny, darmowy)

# ── HTML/XML parsing ──────────────────────────────────────────────────────
beautifulsoup4>=4.12.0      # HTML/XML → Markdown (lokalny, darmowy)
lxml>=5.2.0                 # szybki parser HTML/XML dla BeautifulSoup

# ── Audio / video transcription (KROK 8) ─────────────────────────────────
faster-whisper>=1.0.0       # Local Whisper (CTranslate2, CPU int8, zero cost)
replicate>=0.30.0           # Replicate Whisper large-v3 fallback ($0.0002/min)

# ── HuggingFace Hub — auto-upload dataset i modelu ────────────────────────
huggingface_hub>=0.23.0     # upload datasetu i modelu do HF Hub
datasets>=2.20.0            # HF datasets format (opcjonalny, dla konwersji)

# ── HTTP / async ──────────────────────────────────────────────────────────
httpx>=0.27.0
aiofiles>=23.2.1

# ── Data / numerics ───────────────────────────────────────────────────────
numpy>=1.26.0,<3.0

# ── CLI / progress ────────────────────────────────────────────────────────
tqdm>=4.66.4
rich>=13.7.1
typer>=0.12.3

# ── API / UI ──────────────────────────────────────────────────────────────
fastapi>=0.111.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
python-multipart>=0.0.9
websockets>=12.0,<14.0
streamlit>=1.36.0,<2.0
requests>=2.32.0,<3.0

# ── Gap Scout ─────────────────────────────────────────────────────────────────
apscheduler>=3.10.0,<4.0    # Hourly background scout scheduling
feedparser>=6.0.0            # RSS/Atom parsing for EUR-Lex + arXiv crawlers
trafilatura>=1.6.0           # Article extraction from HTML (boilerplate removal)
pdfminer.six>=20221105       # PDF text extraction (primary; pymupdf is fallback)
openpyxl>=3.1.0              # XLSX extraction → JSON schema for scout ingestion
python-pptx>=0.6.23          # PPTX extraction → slide text + speaker notes
langdetect>=1.0.9            # Language detection ensemble (Step 4 multilingual)
simhash>=2.1.2               # SimHash fingerprinting for content deduplication

# ── Phase 1: Distributed infrastructure ──────────────────────────────────────
temporalio>=1.6.0            # Durable workflow orchestration
nats-py>=2.6.0               # NATS JetStream message bus
litellm>=1.40.0              # Unified LLM gateway (proxy client)

# ── Phase 1: Observability ────────────────────────────────────────────────────
structlog>=24.2.0            # Structured JSON logging
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-sqlalchemy>=0.46b0
opentelemetry-instrumentation-httpx>=0.46b0

# ── Phase 2: Vector memory + agents ──────────────────────────────────────────
qdrant-client>=1.9.0         # Vector DB for long-term Memory Agent
networkx>=3.3                # In-memory graph fallback for KnowledgeGraph

# ── Phase 2: Rate limiting ────────────────────────────────────────────────────
slowapi>=0.1.9               # FastAPI rate limiting

# ── Phase 3: ML experiment tracking ──────────────────────────────────────────
mlflow>=2.14.0               # Experiment tracking for SFT/DPO

# ── Phase 4: Knowledge graph NER ──────────────────────────────────────────────
spacy>=3.7.0                 # Named entity recognition
neo4j>=5.21.0                # Neo4j graph database driver (optional)
